webapp/backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.0
gspread>=6.0
google-auth>=2.0
